Move managers.h into new tests_shared target
