+++ code/memory-mgr/tests/CMakeLists.txt
@@ -1,4 +1,5 @@
 
+add_subdirectory(tests_shared)
 add_subdirectory(unit_tests)
 add_subdirectory(performance_tests)
 add_subdirectory(instantiations)
