+++ code/memory-mgr/tests/tests_shared/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(memory-mgr_tests_shared INTERFACE)
+target_include_directories(memory-mgr_tests_shared INTERFACE
+  ${CMAKE_CURRENT_SOURCE_DIR}/include
+)
