+++ code/memory-mgr/tests/unit_tests/test_allocator.cpp
@@ -21,7 +21,7 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA <http
 Please feel free to contact me via e-mail: shikin@users.sourceforge.net
 */
 
-#include "managers.h"
+#include "tests_shared/managers.h"
 
 #include "memory-mgr/allocator.h"
 #include "memory-mgr/heap_segment.h"
